kernel: fix more lai shit

+++ Kernel/inc/libk.h
@@ -6,7 +6,7 @@
 
 extern bootproto_handoff_t *g_handoff;
 
-void *memset(void *dest, int ch, uint64_t count);
+void *memset(void *dest, int ch, size_t count);
 void *memcpy(void *dest, const void *src, size_t n);
 int memcmp(const void *s1, const void *s2, size_t n);
