+++ Kernel/src/libk.c
@@ -1,7 +1,7 @@
 #include "stdint.h"
 #include "libk.h"
 
-void *memset(void *dest, int ch, uint64_t count) {
+void *memset(void *dest, int ch, size_t count) {
     uint8_t *ptr = dest;
     uint8_t val = ch;
     while (count--) {
